Refactor out common parts of the verifier and quick checker.
